Add Game class
This will be the main interface between generic engine code and game-specific
code.

A game binary is expected to include a single Game-derived class. This is
found during early startup using the object system and instantiated.

For now it just has an Init() function to do game-specific initialisation,
and a GetName() function which is used as the application name (for the main
window title etc.)