Source tree and build simplification
Get rid of separate source and include directories for each component,
and simplify the build system so that it no longer does the automatically
generated include directory in the build tree, and instead just points
directly to the source tree. Also stop SCons from duplicating sources
into the build tree.

This is mainly done to make Visual Studio Code work a bit better - the
old include directory behaviour confused it a lot. Also, duplicating
sources to the build tree and using them to build meant that when
debugging it would not match up the debugged source files to the original
ones.