Can now clear the main window
This commit adds the remaining infrastructure to get to the point of having
the main loop clear the main window. This includes the clear itself, and also
resource barriers.

Resource barriers are exposed directly from the GPU API layer, with no
tracking of resource states internally in that layer. It will be up to higher
level systems to handle this: for frame rendering resources, this will be
done by a frame graph system, while other asset classes will need to manually
perform barriers as needed.

Barriers are implemented using a simplified resource state abstraction over
the Vulkan synchronisation model. This is much closer to D3D12, but with a
bit more fine-grained states.