Initial shader compiler implementation
This is the backend shader compiler. The standard shading language for the
engine will be GLSL using Vulkan semantics. The compiler compiles from this
to SPIR-V. Any other target APIs will need to have an additional translation
step from SPIR-V to their appropriate language (probably using SPIRV-Cross).

Shader compilation is done entirely at runtime for now. In future if we add
a system for baking data to go into a final game build, then we will store
only the compiled SPIR-V, and the game binary would not include the shader
compiler.