Query extensions provided by layers
Some extensions (e.g. VK_EXT_debug_marker) are usually only provided by
layers, we need to query the enabled layers separately to get these.

Also refactor device selection a bit. It now picks the first discrete GPU
available, or the first device in the list if there isn't a discrete GPU.
This is less robust than the code previously was (tries to pick a device that
satisfies all of our requirements), but this allows the code to be cleaner.
Could add some user-configurable device selection mechanism in future.