Handle shader includes ourself rather than with glslang
The reason for this is that glslang does not substitute included source text
into the source embedded in SPIR-V modules' OpSource (when debug info is
enabled). This means that we don't see the full shader source in a RenderDoc
capture, and therefore cannot edit shaders from there without manually
setting up include paths every time.

As part of this, we now look in the shader search paths for source files if
the included path cannot be found relative to the path of the current source
file.