Use HLSL semantics for vertex attribute matching
GPUVertexInputState now specifies a semantic and index rather than a raw
location, which is matched to a raw location using reflection information
from the SPIR-V shader binary.

This is being done because it'll make it much easier to pair arbitary model
data and shaders - we have semantic information to pair attributes from the
model with what the shader wants.