Refactor how swapchain texture views work
Rather than creating a single view internally and mandating that only that
view can be used to access the swapchain texture, allow arbitrary views to
be created. However, this has the restriction that views to the swapchain
can only exist between BeginPresent()/EndPresent(), so that they refer to
the currently acquired image for the swapchain.

This is a better fit for how things will work with the render graph.