Create views for render graph resources, call begin/end callbacks
This moves the view array and GetView() method onto the pass, and makes
ResourceViewHandle specific to a pass. The reason for this is that we need to
create views at the time that we execute each pass, after the begin/end
callbacks for imported resources, due to the rules about creating views of
swapchain textures (and we only want to call those callbacks tightly around
execution of the passes that refer to those resources, so that the Vulkan
swapchain acquire gets done as late in the frame as possible).