Require explicit code generation for enums
Before, any file which we generate code for would include code for any enum
marked with ENUM() in any included files. These are deduplicated at link-time,
however it still results in a lot of extra generated code that needs to be
compiled for enums defined in include files that are used in a lot of places.

Change the behaviour to match what we do for classes: only generate code for
enums marked with ENUM() in the file that we're generating code for.