Use the material system to draw entities
Replacing the initial hack in BasicRenderPipeline to use the same shader
for all entities and look up PSOs at draw time.

PSOs for entities now get created at the time the Renderer component gets
activated in the world. This means we don't have to look up PSOs at draw
time. This may need some rework in future so that we can create them at
world load time (or stream them asynchronously), but this is good enough
for now.