Work in progress glTF importer
This is different to a normal asset loader: asset loaders are for loading a
single asset, whereas glTF files contain data for potentially many different
assets and entities that need to be composed together.

This importer is intended to be used as a one-time process to import/convert
a glTF file to a set of assets/entities usable by the engine. It will create
and save assets for all the textures/meshes/materials in the file, and then
create entities in a world composing the scene.

So far, this just loads in everything from the file and validates it. Next
part will be to generate engine assets/entities out of all of it.