Finish asset/world generation for the glTF importer
This is enough to load the glTF DamagedHelmet sample. TestGame is now using
a world that was generated from DamagedHelmet.gltf and serialised - running
that will load the generated world rather than re-importing the glTF file.
The code used to import is still in TestGame.cpp, commented out.

For now, we're just mapping glTF materials to the BasicTextured shader using
the model's base colour texture. PBR implementation will come soon (the whole
point of writing this importer is to give me an easy way to get PBR content
into the engine).