Fix incorrect swapchain image layout transition before ImGUI rendering
VulkanContext::ResourceBarrier() was always using UNDEFINED as the old Vulkan
layout when transitioning away from the Present state.

Since ImGUI was moved outside the render graph, the render graph transitions
to Present when it's done, then ImGUI transitions from Present to RenderTarget
if it has anything to render. The Vulkan backend would use UNDEFINED as the
old layout here, which can corrupt what was rendered to the swapchain image by
the render graph.

Fix by tracking whether a transition on a swapchain image is the first in the
frame, and only use UNDEFINED on first use.