Add a function to prevent the render graph culling a pass
Passes whose outputs are not consumed by any pass will normally be skipped,
but this can get in the way during development where the pass that consumes a
pass's outputs has not yet been implemented. This function can be used in this
case to force a pass to be executed.