Support debug output of RenderGraph texture resources
This allows specific versions of texture resources to be selected in the
RenderGraph UI as debug output. After the pass which produces the selected
version of the resource, it will be copied out into an intermediate texture.
At the end of graph execution it will be blitted onto the RenderOutput that
the resource was created within.