Ignore special enum constants that aren't valid values
For some enums we have a special constant indicating the maximum value of the
enum, used in the code for loops, array sizes, etc. By convention these are
named as kEnumNameCount (for an enum named EnumName). These are not actually
valid values for the enum to be, so ignore these when generating code for the
enum. This prevents these count constants from being presented as a choice in
property editor UIs.