Support pragma once in shaders
Our include handling would just paste a copy of the included file into the
monolithic source we pass to shaderc. If a header is included multiple times
it would appear multiple times in that source, which both bloats the source
(possibly increasing compile time), and could lead to confusion when editing
shaders in RenderDoc.

To avoid this, support pragma once and use this everywhere: when we encounter
it and the source is already included, don't add in another copy.