Add feature flags for emissive and occlusion to the PBR shader
These now use different shader variants. When the features are disabled we
don't sample textures for them (previously, to disable them, dummy textures
had to be bound such that the value sampled removes the emissive/occlusion
factors). Also set the DeferredOpaque_Emissive flag so that emissive output
is disabled on materials that don't need it.