Bind a zero buffer when a required vertex input is not supplied
When a vertex shader uses an input which we do not have a match for in the
pipeline vertex input state, bind a dummy zero-filled buffer to that input
whenever the pipeline is used, so that it will read as zero (stride is set
to 0, so we only need to supply 1 zero value regardless of vertex count).

I'm adding this as a workaround for a glTF model I have which has a few
meshes that don't supply texture coordinates but the material does have
textures. I can't find anything in the spec to say what should happen in
this case, but the official glTF sample viewer will use 0 for the texture
coordinates.

The alternative to this would be to generate shader variants based on what
vertex attributes the mesh has, but this is quite a bit more work and I don't
want to do this until I have more need for it.